+++ src/lib.ts
@@ -2,6 +2,7 @@ export { Decimal } from 'decimal.js';
 export { CharString } from './CharString';
 export { ComplexDecimal } from './ComplexDecimal';
 export { MultiArray } from './MultiArray';
+export { Structure } from './Structure';
 export { LinearAlgebra } from './LinearAlgebra';
 export { MathOperation } from './MathOperation';
 export { CoreFunctions } from './CoreFunctions';